ipfs: Stop using the stat API
To reduce the total number of calls to IPFS, and
to generally simplify our interface on IPFS, this
stops using the `stat` API to check existence and
control file size. Both can be achieved efficiently
with the more usual cat api.

Relying on files/stat was questionable anyways,
since we don't know to what extent IPFS clients
actually enforce that the declared metadata size value
matches the real file size.

+++ node/src/manager/commands/run.rs
@@ -73,7 +73,7 @@ pub async fn run(
     let ipfs_client = ipfs_clients.first().cloned().expect("Missing IPFS client");
     let ipfs_service = ipfs_service(
         ipfs_client,
-        env_vars.mappings.max_ipfs_file_bytes as u64,
+        env_vars.mappings.max_ipfs_file_bytes,
         env_vars.mappings.ipfs_timeout,
         env_vars.mappings.ipfs_request_limit,
     );
